style: Fix import list style and split parser module internals
